organize DataFrame tests
